orchestrator/requirements.txt
# Core dependencies
docker==7.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# AI APIs
anthropic==0.39.0
openai==1.57.2

# Utilities
requests==2.31.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
